chore: move mocks folder
